Create top level start.py. Train Bayes classifier.
The top level start.py parses arguments to see
which data source it should use and which ML
algorithm it should use. Only Slate and Bayes
are supported right now.

We might need to tweak the provider interface a
little bit to make start.py support more general
types of algorithms. But, this is a good start.